# ferrox-cp/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ferrox Control Plane API",
    "description": "Control plane for the Ferrox LLM gateway: API client administration, usage and budgets, signing-key rotation and the audit log (admin key required), plus the public token exchange, JWKS and health endpoints.",
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/.well-known/jwks.json": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "`GET /.well-known/jwks.json`",
        "description": "Returns all active RSA public keys in JWKS format (RFC 7517).\nNo authentication required — this endpoint is called by the gateway.\nResponse is cacheable for 5 minutes via `Cache-Control: max-age=300`.",
        "operationId": "jwks_handler",
        "responses": {
          "200": {
            "description": "Active RSA public keys (RFC 7517 JWKS), cacheable for 5 minutes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JwksResponse"
                }
              }
            }
          },
          "500": {
            "description": "Keys could not be loaded; empty key set, not cacheable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JwksResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/audit": {
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "`GET /api/audit`",
        "description": "Lists audit log entries with optional filters.\nQuery parameters: `client_id`, `event`, `since`, `limit`, `offset`.",
        "operationId": "list_audit",
        "parameters": [
          {
            "name": "client_id",
            "in": "query",
            "description": "Only entries for this client.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "nullable": true
            }
          },
          {
            "name": "event",
            "in": "query",
            "description": "Only entries of this event type (e.g. `token_issued`).",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "since",
            "in": "query",
            "description": "Only entries created at or after this time.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size (capped at 1000).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 100,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of entries to skip.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit log entries, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEntry"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_auth": []
          }
        ]
      }
    },
    "/api/clients": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "`GET /api/clients`",
        "operationId": "list_clients",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Page size (capped at 1000).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 50,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of records to skip.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Clients, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "`POST /api/clients`",
        "description": "Creates a new API client.  Generates an `sk-cp-` prefixed key, hashes it\nwith bcrypt, stores the hash, and returns the plaintext key once in the\nresponse body.",
        "operationId": "create_client",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClientRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Client created; `api_key` is returned only in this response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateClientResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "A client with this name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database or key-hashing failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_auth": []
          }
        ]
      }
    },
    "/api/clients/{id}": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "`GET /api/clients/:id`",
        "operationId": "get_client",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Client id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Client not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Clients"
        ],
        "summary": "`DELETE /api/clients/:id`",
        "operationId": "revoke_client",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Client id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Client revoked"
          },
          "401": {
            "description": "Missing or invalid admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Client not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_auth": []
          }
        ]
      }
    },
    "/api/clients/{id}/budget": {
      "patch": {
        "tags": [
          "Clients"
        ],
        "summary": "`PATCH /api/clients/:id/budget`",
        "description": "Update token budget settings for a client.",
        "operationId": "update_client_budget",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Client id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBudgetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Client not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_auth": []
          }
        ]
      }
    },
    "/api/clients/{id}/reactivate": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "`POST /api/clients/:id/reactivate`",
        "description": "Re-activate a revoked client and reset its budget period.",
        "operationId": "reactivate_client",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Client id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Client reactivated"
          },
          "401": {
            "description": "Missing or invalid admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Client not found or already active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_auth": []
          }
        ]
      }
    },
    "/api/clients/{id}/usage": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "`GET /api/clients/:id/usage`",
        "description": "Returns aggregated token usage from `usage_log` for the last 24h, 7d, and 30d.",
        "operationId": "client_usage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Client id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token usage over the last 24h, 7d and 30d",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Client not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_auth": []
          }
        ]
      }
    },
...
    "/api/signing-keys": {
      "get": {
        "tags": [
          "Signing keys"
        ],
        "summary": "`GET /api/signing-keys`",
        "description": "Lists all signing keys (active and retired).  Private key material is never\nincluded in the response.",
        "operationId": "list_signing_keys",
        "responses": {
          "200": {
            "description": "All signing keys, active and retired",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SigningKeyResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_auth": []
          }
        ]
      }
    },
    "/api/signing-keys/rotate": {
      "post": {
        "tags": [
          "Signing keys"
        ],
        "summary": "`POST /api/signing-keys/rotate`",
        "description": "Generates a new RSA-2048 keypair and schedules the current active key(s)\nfor retirement after a grace period equal to the longest client token TTL.\nBoth keys remain in the JWKS during the overlap window so in-flight tokens\nstay verifiable.",
        "operationId": "rotate_keys",
        "responses": {
          "201": {
            "description": "The newly generated signing key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SigningKeyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Key generation, configuration or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditEntry": {
        "type": "object",
        "description": "A single entry in the immutable audit log.",
        "required": [
          "id",
          "event",
          "created_at"
        ],
        "properties": {
          "client_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "event": {
            "type": "string",
            "description": "Event type: `token_issued`, `client_created`, `client_revoked`,\n`key_rotated`, `budget_exceeded`, or any other recorded string.",
            "example": "token_issued"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "metadata": {
            "type": "object",
            "description": "Free-form, event-specific JSON details.",
            "nullable": true
          }
        }
      },
...
      "JwksResponse": {
        "type": "object",
        "required": [
          "keys"
        ],
        "properties": {
          "keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Jwk"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "admin_auth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The static admin key configured as `CP_ADMIN_KEY`."
      },
      "client_key_auth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A client API key (`sk-cp-...`) as returned by `POST /api/clients`."
      }
    }
  },
...
}